requirements.txt
requests
dotenv
streamlit
pandas
pytesseract
Pillow
spacy